src/main/java/com/favouritefilms/services/UserService.java
package com.favouritefilms.services;

import com.favouritefilms.repositories.UserRepo;
import com.favouritefilms.entities.User;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class UserService
        implements UserDetailsService {

    private final UserRepo userRepo;
    private final PasswordEncoder passwordEncoder;

    @Autowired
    public UserService(UserRepo userRepo, PasswordEncoder passwordEncoder) {
        this.userRepo = userRepo;
        this.passwordEncoder = passwordEncoder;
    }

    @Override
    public UserDetails loadUserByUsername(String username)
            throws UsernameNotFoundException {
        User user = userRepo.findByUsername(username);
        if (user != null) {
            return user;
        }
        throw new UsernameNotFoundException(
                "User '" + username + "' not found");
    }

    public boolean save(User newUser) {

        User userFromDb = userRepo.findByUsername(newUser.getUsername());

        if(userFromDb != null){
            return false;
        }

        newUser.setPassword(passwordEncoder.encode(newUser.getPassword()));
        userRepo.save(newUser);

        return true;
    }
}
